// src/app/domain/type/kind/technology-category-kind-domain.type.ts
export type TechnologyCategoryKindDomainType =
  | 'markupLanguage'
  | 'stylesheetLanguage'
  | 'programmingLanguage'
  | 'framework'
  | 'library'
  | 'test'
  | 'codeQuality'
  | 'database'
  | 'server'
  | 'runtime'
  | 'containerization'
  | 'integrationPlatform'
  | 'api'
  | 'versionControlSystem'
  | 'packageManager'
  | 'monorepo'
  | 'ide'
  | 'editor'
  | 'diagram'
  | 'projectManagement'
  | 'operatingSystem';